git-svn-id: http://fussvn.fusion.ccfe.ac.uk/svnroot/EPICS@619 9c91d93d-f138-4f03-b554-451257969c6d

+++ LeyboldTurbo/LeyboldSimApp/src/LeyboldSimPortDriver.h
@@ -23,7 +23,7 @@ protected:
 private:
 	std::map<std::string, int> m_Parameters;
 	epicsMutexId m_mutexId;
-	size_t m_NumConnected;
+	int m_NumConnected;
 	volatile bool m_Exiting;
 };
